Refactor everything into Table component
